+++ js_modules/dagster-ui/packages/ui-core/src/launchpad/LaunchpadSession.tsx
@@ -19,7 +19,6 @@ import {
   TextInput,
   isHelpContextEqual,
 } from '@dagster-io/ui-components';
-import merge from 'deepmerge';
 import uniqBy from 'lodash/uniqBy';
 import * as React from 'react';
 import styled from 'styled-components';
@@ -53,6 +52,7 @@ import {
   PreviewConfigQuery,
   PreviewConfigQueryVariables,
 } from './types/LaunchpadSession.types';
+import {mergeYaml, sanitizeConfigYamlString} from './yamlUtils';
 import {showCustomAlert} from '../app/CustomAlertProvider';
 import {
   IExecutionSession,
@@ -875,8 +875,6 @@ const deletePropertyPath = (obj: any, path: string) => {
   }
 };
 
-const sanitizeConfigYamlString = (yamlString: string) => (yamlString || '').trim() || '{}';
-
 const PREVIEW_CONFIG_QUERY = gql`
   query PreviewConfigQuery(
     $pipeline: PipelineSelector!
@@ -919,45 +917,3 @@ const PIPELINE_EXECUTION_CONFIG_SCHEMA_QUERY = gql`
 
   ${CONFIG_EDITOR_RUN_CONFIG_SCHEMA_FRAGMENT}
 `;
-
-/**
- * Utility function to merge two YAML documents:
- * - Stringify string values with quotes. Avoids known issues with "5:30" becoming 5:30 which parses as a Sexagesimal number
- * - When merging arrays, combine rather than concatenate (the default deepmerge behavior)
- * -
- */
-function mergeYaml(
-  base: string | Record<string, any>,
-  overrides: string | Record<string, any>,
-  extraOpts?: yaml.SchemaOptions,
-) {
-  const baseObj = typeof base === 'string' ? yaml.parse(sanitizeConfigYamlString(base)) : base;
-
-  const overridesObj =
-    typeof overrides === 'string' ? yaml.parse(sanitizeConfigYamlString(overrides)) : overrides;
-
-  const arrayCombineMerge: merge.Options['arrayMerge'] = (target, source, options) => {
-    const destination = target.slice();
-
-    source.forEach((item, index) => {
-      if (typeof destination[index] === 'undefined') {
-        destination[index] = options?.cloneUnlessOtherwiseSpecified(item, options);
-      } else if (options?.isMergeableObject(item)) {
-        destination[index] = merge(target[index], item, options);
-      } else if (target.indexOf(item) === -1) {
-        destination.push(item);
-      }
-    });
-    return destination;
-  };
-
-  const mergedObj = merge(baseObj, overridesObj, {
-    arrayMerge: arrayCombineMerge,
-  });
-
-  return yaml.stringify(mergedObj, {
-    defaultKeyType: 'PLAIN',
-    defaultStringType: 'QUOTE_SINGLE',
-    ...extraOpts,
-  });
-}
